+++ Pandora/Pattern/Functor/Monoidal.hs
@@ -1,0 +1,5 @@
+module Pandora.Pattern.Functor.Monoidal where
+
+import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal)
+
+type family Unit (p :: * -> * -> *) :: r